A little cleaned up. tact.jl is almost unused. But may be referenced for proof struct. testdvcreso.jl is not enough. When a big complex data is used in a funtcion, the test of it is tedious. How can I resolve it.
	modified:   README
	modified:   coreprint.jl
	modified:   dvcreso.jl
	modified:   play.jl
	modified:   primitives.jl
	modified:   tact.jl
	modified:   testall.jl
	modified:   testdvcreso.jl
