+++ Resolution/primitives.jl
@@ -1,5 +1,13 @@
 # primitives.jl
 
+#==
+ Key concept is GOAL
+ but not define it as type.
+
+ GOAL is a literal Array.
+ And a Set of GOAL should be maintained.
+==#
+
 ## Type declare
 
 FTerm=Expr
@@ -9,6 +17,14 @@ Const = Union{Symbol, Number}
 
 Vlist = Array
 Tlist = Array
+
+struct EQTerm
+ sign
+ psym
+ body
+end
+
+Equation=Array{EQTerm}
 
 ## primitives
 isvar(sym::Symbol, vars::Vlist)=sym in vars
