+++ Resolution/testall.jl
@@ -1,8 +1,5 @@
 include("testreso.jl")
-#include("testcgop.jl")
-#include("testtact.jl")
 include("testutils.jl")
-#include("setdata.jl")
 include("testnewcore.jl")
 include("testdvcreso.jl")
